pyproject.toml
[project]
name = "WorkerAnt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "fastapi[standard]>=0.135.3",
    "flask>=3.1.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
